Support many routes in fuse
