+++ services/core/fuse/envoy-xds-config.yaml
@@ -7,7 +7,7 @@ admin:
   access_log_path: /dev/null
   address:
     socket_address:
-      address: 127.0.0.1
+      address: 0.0.0.0
       port_value: 19000
 
 dynamic_resources:
@@ -41,7 +41,7 @@ static_resources:
             address:
               socket_address:
                 # address in which fuse is running on
-                address: 127.0.0.1
+                address: 0.0.0.0
                 port_value: 18000
     http2_protocol_options: {}
     name: xds_cluster
