+++ services/examples/file_host/config.yaml
@@ -19,17 +19,19 @@ service:
   network:
     # the bind address when opening ports for the web server
     # default: "0.0.0.0"
-    bind_address: "localhost"
+    bind_address: "0.0.0.0"
     # the port the process will bind to (if it hosts a web server of any kind)
     # default: 8090
     port: 8080
     # routing information within the draft cluster
     internal:
       # e.g. this could be http://file_host.examples.svc.cluster.local:8080 if running in kubernetes
-      host: localhost
+      host: localhost 
       # the port to route internal traffic to (e.g. envoy -> process)
       port: 8080
     # routing information from outside of the draft cluster
     external:
+      # The downstream host the request if coming from. This is used in the domains
+      # configuration of the `RouteConfiguration` virtual host. Also known as domain routing.
       # e.g. this could be file_host.examples.steady-bytes.com
-      host: localhost
+      host: localhost
